Fix for ConcurrentModificationError on WrappedInUseEntityPacket
This fix works by offloading the for loop, which finds the correct entity, onto the main thread. I believe this ConcurrentModificationError occured due to the list of entities grabbed by the Bukkit API is being written to when the entities are being updated every tick.

This error would randomly pop up and is for the most part, harmles. However, it can be quite annoying for users. That being so, I set out to fixing it.