Fixes bug with BlockBox system
There was aan Asychronous Chunk Loading error running with the boundingbox part of this system. This would be caused by running it asynchrounsly, off the main thread. I added a check which will ensure that the chunk of where the bounding box is being checked is loaded before running it asynchronously.